refactor: add home private folder
